ESP32UCSG/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==42.0.0
python-dotenv==1.0.0
pydantic==2.5.0